Move einsum into einsum.py
